feat: Handling the files upload
Adding an api view for setting up the items pictures
the upload image is saved in the image folder, but
code is already written to save them in an s3 Bucket.
For now that code is commented since it is not yet tested.
The only thing missing for now is updating the Item's
image_url value in the database.